requirements.txt
anyio==3.6.2
anytree==2.8.0
apparmor==3.1.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciidoc==10.2.0
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
autocommand==2.2.2
Babel==2.11.0
backcall==0.2.0
bauh==0.10.4
beautifulsoup4==4.11.1
bleach==6.0.0
btrfsutil==6.2.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.2
cssselect==1.2.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
distro==1.8.0
Django==4.1.3
djangorestframework==3.14.0
docutils==0.19
entrypoints==0.4
evdev==1.6.1
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fqdn==1.5.1
future==0.18.2
Glances==3.3.0
gunicorn==20.1.0
idna==3.4
inflect==6.0.2
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_packaging==0.12.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.19.0
lensfun==0.3.3
LibAppArmor==3.1.3
libtorrent==2.0.8
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
meson==1.0.1
mistune==2.0.4
more-itertools==9.0.0
multidict==6.0.3
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nftables==0.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
ordered-set==4.1.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pycairo==1.23.0
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
PyGObject==3.42.2
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
Reflector==2021.11.20.2.41.3
requests==2.28.1
requests-toolbelt==0.10.1
requests-unixsocket==0.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
Send2Trash==1.8.0
setproctitle==1.3.2
shtab==1.5.8
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
team==1.0
termcolor==2.2.0
terminado==0.17.1
testpath==0.6.0
tinycss2==1.2.1
tldr==3.1.0
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.1
trove-classifiers==2023.3.1
typing_extensions==4.5.0
uc-micro-py==1.0.1
uri-template==1.2.0
urllib3==1.26.12
validate-pyproject==0.12.1
virtualenv==20.16.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
